% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_cov}
\alias{exponential_cov}
\title{exponential_cov func}
\usage{
exponential_cov(x, y, params)
}
\arguments{
\item{x}{vec}

\item{y}{vec}

\item{params}{list for theta}
}
\description{
calculate the exponential covariance function
}
