% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gaussian_kernel}
\alias{Gaussian_kernel}
\title{Gaussian_kernel func}
\usage{
Gaussian_kernel(x, y, kernel_width)
}
\arguments{
\item{x}{mat X(m, p)}

\item{y}{mat Y(n, p)}

\item{kernel_width}{double}
}
\description{
Calculate the Gaussian_kernel function using RcppArmadillo
}
\examples{
a = matrix(c(1, 2, 3, 4), nrow=2)
b = matrix(c(2, 3, 4, 5), nrow=2)
Gaussian_kernel(a, b, kernel_width = 1.0)
}
