% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{Gaussian_reg_fit_pre}
\alias{Gaussian_reg_fit_pre}
\title{One dimensional gaussian process regression}
\usage{
Gaussian_reg_fit_pre(
  X,
  y,
  Xtest,
  alpha = 1e-10,
  kernel = exponential_cov,
  theta,
  plot_f = FALSE
)
}
\arguments{
\item{X}{input matrix}

\item{y}{input value}

\item{Xtest}{predict matrix}

\item{alpha}{error parameter}

\item{kernel}{kernel used in gaussian process, by default we use exponential cov}

\item{theta}{parameters for kernel}

\item{plot_f}{plot f_prior or not}
}
\value{
a list containing posterior mean, std, posterior sample and marginal likelihood
}
\description{
naive implementation for gaussian process regression
}
\examples{
x <- seq(-20,20,0.1)
y <- sin(x)/x + rnorm(401,sd=0.03)
res = Gaussian_reg_fit_pre(x, y , x, alpha=1e-5, theta=list(1, 1/(3.88)^2), plot_f=TRUE)
plot(x, y, type ="l")
lines(x, res$mu, col="red")

}
