% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{run}
\alias{run}
\title{main function running for spatiotemporal prediction}
\usage{
run(Xtrain_s, Xtrain_t, ytrain, Xtest_s, Xtest_t, kernel, parameters)
}
\arguments{
\item{Xtrain_s}{spatial train matrix}

\item{Xtrain_t}{temporal train matrix}

\item{ytrain}{ytrain matrix}

\item{Xtest_s}{spatial test matrix}

\item{Xtest_t}{temporal test matrix}

\item{kernel}{kernel, default: Gaussian kernel}

\item{parameters}{param for kernel,spatial and temporal}
}
\value{
prediction list containing mean and variance
}
\description{
main function running for spatiotemporal prediction
}
