% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{spatiotemporal_gpr_train}
\alias{spatiotemporal_gpr_train}
\title{training function for spatiotemporal gpr}
\usage{
spatiotemporal_gpr_train(K_1_tra_tra, K_2_tra_tra, Y_tra, parameters)
}
\arguments{
\item{K_1_tra_tra}{kernel for train}

\item{K_2_tra_tra}{kernel for train}

\item{Y_tra}{train y}

\item{parameters}{params}
}
\value{
state information
}
\description{
training function for spatiotemporal gpr
}
