% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gaussian_kernel}
\alias{Gaussian_kernel}
\title{Gaussian_kernel func}
\usage{
Gaussian_kernel(x, y, kernel_width)
}
\arguments{
\item{x}{mat}

\item{y}{mat}

\item{kernel_width}{double}
}
\description{
Calculate the Gaussian_kernel function
}
