% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{Gaussian_plot_pos}
\alias{Gaussian_plot_pos}
\title{Gaussian_plot_pos}
\usage{
Gaussian_plot_pos(Xtest, Xtrain, ytrain)
}
\arguments{
\item{Xtest}{input test matrix, same as function Gaussian_reg_fit_pre}

\item{Xtrain}{input training matrix X, same as function Gaussian_reg_fit_pre}

\item{ytrain}{input training vector y, same as function Gaussian_reg_fit_pre}
}
\value{
plot func, no return variable
}
\description{
Plot using result from Gaussian process
}
\details{
plot result after gaussian process regression, including measuring uncertainty
}
\examples{
Xtest = seq(-5, 5, by=10/49)
theta = list(1, 10)
Xtrain = c(-4, -3, -2, -1, 1)
y = sin(Xtrain)
Gaussian_plot_pos(Xtest, Xtrain, y)
}
