% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{Gaussian_reg_fit_pre}
\alias{Gaussian_reg_fit_pre}
\title{Gaussian_reg_fit_pre}
\usage{
Gaussian_reg_fit_pre(
  X,
  y,
  Xtest,
  alpha = 1e-10,
  kernel = exponential_cov,
  theta,
  plot_f = FALSE
)
}
\arguments{
\item{X}{Input matrix for training}

\item{y}{Input value y}

\item{Xtest}{Predict matrix}

\item{alpha}{error parameter, \strong{default:1e-10}}

\item{kernel}{Kernel used in gaussian process, by default we use \strong{exponential cov} i.e. RBF kernel}

\item{theta}{parameters for kernel function}

\item{plot_f}{whether plot f_prior or not}
}
\value{
a list containing posterior mean, std, posterior sample and marginal likelihood
}
\description{
One dimensional Gaussian Process Regression
}
\details{
Naiiive implementation for one dimensional gaussian process regression, which first calculates prior, then uses bayes rules for posterior
}
\examples{
x <- seq(-20,20,0.1)
y <- sin(x)/x + rnorm(401,sd=0.03)
res = Gaussian_reg_fit_pre(x, y , x, alpha=1e-5, theta=list(1, 1/(3.88)^2), plot_f=TRUE)
plot(x, y, type ="l")
lines(x, res$mu, col="red")

}
\seealso{
\code{\link{Gaussian_plot_pos}}
}
