% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{spatiotemporal_gpr_test}
\alias{spatiotemporal_gpr_test}
\title{fit function for spatiotemporal gpr}
\usage{
spatiotemporal_gpr_test(
  K_1_test_tra,
  K_1_test_test,
  K_2_test_tra,
  K_2_test_test,
  state
)
}
\arguments{
\item{K_1_test_tra}{kernel for train_test}

\item{K_1_test_test}{kernel for test}

\item{K_2_test_tra}{kernel for train_test}

\item{K_2_test_test}{kernel for test}

\item{state}{state return from train func}
}
\value{
prediction
}
\description{
fit function for spatiotemporal gpr
}
