% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_process.R
\name{spatiotemporal_gpr_train}
\alias{spatiotemporal_gpr_train}
\title{training function for spatiotemporal gpr}
\usage{
spatiotemporal_gpr_train(K_1_tra_tra, K_2_tra_tra, Y_tra, parameters)
}
\arguments{
\item{K_1_tra_tra}{kernel for spatial train data}

\item{K_2_tra_tra}{kernel for temporal train data}

\item{Y_tra}{train y}

\item{parameters}{params for spatial kernel, temporal kernel and the sigma. \strong{Suggest:}The kernel width parameters are chosen as the means of pairwise Euclidean distances between training instances for all kernels.}
}
\value{
State information, which will be used in fit function
}
\description{
training function for spatiotemporal gpr
}
